+++ frontend/src/App.js
@@ -7,6 +7,7 @@ import GlobalStyle from "./components/GlobalStyle"
 
 import { BrowserRouter, Routes, Route, useLocation } from "react-router-dom";
 import { Home, CryptidList, Cryptid } from "./pages";
+import HomeContainer from './containers/HomeContainer';
 import { Header, Footer } from "./components/layouts";
 import { PageContainer } from "./components/layouts";
 import ScrollToTop from './common/ScrollToTop'
@@ -36,7 +37,7 @@ const App = () => {
         <PageContainer>
           <ScrollToTop />
           <Routes>
-            <Route path="/" element={<Home />} />
+            <Route path="/" element={<HomeContainer />} />
             <Route path="/cryptids" element={<CryptidList />} />
             <Route path="/cryptids/:id" element={<Cryptid />} />
           </Routes>
