+++ frontend/src/pages/index.js
@@ -1,3 +1,3 @@
-export { default as Home } from "./Home/Home";
+export { default as HomePage } from "./Home/HomePage";
 export { default as CryptidList } from "./CryptidList";
 export { default as Cryptid } from "./Cryptid";
